Layout: use selector for layoutFocus rather than observer
The observer requires that the Flux store emit an event to update the component,
but if anything uses the Redux store to change the layoutFocus, an event will
not be emitted.

Instead, this uses the `getCurrentLayoutFocus` selector in `mapStateToProps` so
that when the Redux store changes, the component will update.