Generate site titles dynamically
Previous work enabled editing the format string for site titles on
various types of page views: front page, post, page, etc...

However, none of the previews that have been being added display that
title.

This PR introduces the functionality necesary to generate that title
dynamically. This is important because it can be too slow to hear back
from the API if we were to try and save/load while editing it.