Editor: Add success notice for scheduled posts
We display success notices for published and updated posts. This adds
the same thing to scheduled posts along with a preview link. It should
not display when scheduled posts are autosaved. To match the current
behavior for published posts, the "Save" button also doesn't appear when
a post is scheduled.