Fixed Desktop bug where navigating would result in white flash.
Currently on the Desktop app clicking on any link causes the the view
about to be loaded to render twice resulting in an ugly white flash.
Here’s a video - https://cloudup.com/cSs5t4skbcm.

The cause of this bug is actually two fold.
https://github.com/Automattic/wp-calypso/pull/7121 is what caused the
bug to come to the surface because the code as initially written didn’t
properly execute for the desktop but the source of this bug lies in
faulty logic that was introduced in
https://github.com/Automattic/wp-calypso/pull/4506. The problem with the
code referenced in both PRs is that the auth controller checks to see if
we need to redirect to `/login/ or `/authorize` depending on web vs
desktop and if not it calls `next()` which will render the appropriate
page. The bug occurs because if you follow the flow in `checkToken`
`next()` will be called twice.`